change type of column 'value.weight_values'

+++ docs/sql/make_table.sql
@@ -16,7 +16,7 @@ CREATE TABLE IF NOT EXISTS `word_count`(
 CREATE TABLE IF NOT EXISTS `weight_values`(
    `weight_id` SMALLINT(6) AUTO_INCREMENT NOT NULL COMMENt '重みベクトルの添字',
    `keystr`    VARCHAR(255)               NOT NULL COMMENT '重みベクトルのキー',
-   `value`     INT(11)          DEFAULT 1 NOT NULL COMMENt '重みベクトルの値',
+   `value`     FLOAT(18, 17)    DEFAULT 1 NOT NULL COMMENt '重みベクトルの値',
    PRIMARY KEY (`weight_id`)
 ) DEFAULT CHARSET=utf8 COMMENT='重みベクトルの値';
